server: move isFunnelRequest() into server.go
